move `like` function back to core
