+++ src/expr/core/src/scalar/like.rs
