+++ src/expr/core/src/scalar/mod.rs
@@ -12,26 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![allow(non_snake_case)] // for `ctor` generated code
-#![feature(let_chains)]
-#![feature(lint_reasons)]
-#![feature(iterator_try_collect)]
-#![feature(lazy_cell)]
-#![feature(generators)]
-#![feature(arc_unwrap_or_clone)]
-
-extern crate self as risingwave_expr;
-
-pub mod aggregate;
-#[doc(hidden)]
-pub mod codegen;
-mod error;
-pub mod expr;
-pub mod scalar;
-pub mod sig;
-pub mod table_function;
-pub mod window_function;
-
-pub use error::{ExprError, Result};
-pub use risingwave_common::{bail, ensure};
-pub use risingwave_expr_macro::*;
+pub mod like;
